requirements.txt
python-whois
requests
rich